pyproject.toml
[project]
name = "subtitle-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.11.0",
]
